requirements.txt
# Agent framework
langgraph>=0.0.40
langchain>=0.1.0
langchain-google-genai>=0.0.6

# Web scraping
selenium>=4.15.0
beautifulsoup4>=4.12.0
lxml>=4.9.0

# API clients
python-telegram-bot>=20.7
google-generativeai>=0.3.0

# Utilities
pydantic>=2.5.0
python-dotenv>=1.0.0

# Logging
structlog>=23.2.0
